refactor: 핵심 비즈니스 로직(도메인) 관련 파일을 domain 폴더 하위로 이동
- Car.js 이동
- RacingcarManager.js 이동
