+++ src/domain/Car.js
@@ -1,5 +1,6 @@
 const MOVE_CONDITION = 4
 const MOVE_FORWARD = 1;
+
 class Car {
     name
     position
@@ -13,6 +14,6 @@ class Car {
             this.position += MOVE_FORWARD;
         }
     }
-
 }
+
 export default Car;
